Devtask/2 repositoriy employee

+++ internal/service/staffservice/staff_service.go
@@ -7,10 +7,10 @@ import (
 )
 
 type StaffRepo interface {
-	Create(ctx context.Context, u domain.Employee) (int, error)
-	Read(ctx context.Context, id int) (domain.Employee, error)
-	Update(ctx context.Context, id int, u domain.Employee) (domain.Employee, error)
-	Delete(ctx context.Context, id int) error
+	Create(ctx context.Context, u domain.Employee) (uint64, error)
+	Read(ctx context.Context, id uint64) (domain.Employee, error)
+	Update(ctx context.Context, id uint64, u domain.Employee) (domain.Employee, error)
+	Delete(ctx context.Context, id uint64) error
 	ReadAll(ctx context.Context) ([]domain.Employee, error)
 }
